Move API Key Section to Profile Page
